Rename log4j mdc library package name
