+++ instrumentation/logback/logback-mdc-1.0/javaagent/src/test/groovy/io/opentelemetry/javaagent/instrumentation/logback/v1_0/LogbackTest.groovy
@@ -5,7 +5,7 @@
 
 package io.opentelemetry.javaagent.instrumentation.logback.v1_0
 
-import io.opentelemetry.instrumentation.logback.v1_0.AbstractLogbackTest
+import io.opentelemetry.instrumentation.logback.mdc.v1_0.AbstractLogbackTest
 import io.opentelemetry.instrumentation.test.AgentTestTrait
 
 class LogbackTest extends AbstractLogbackTest implements AgentTestTrait {
