+++ instrumentation/logback/logback-mdc-1.0/library/src/main/java/io/opentelemetry/instrumentation/logback/mdc/v1_0/internal/UnionMap.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.logback.v1_0.internal;
+package io.opentelemetry.instrumentation.logback.mdc.v1_0.internal;
 
 import java.util.AbstractMap;
 import java.util.AbstractSet;
