+++ instrumentation/logback/logback-mdc-1.0/library/src/test/groovy/io/opentelemetry/instrumentation/logback/mdc/v1_0/internal/UnionMapTest.groovy
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.logback.v1_0.internal
+package io.opentelemetry.instrumentation.logback.mdc.v1_0.internal
 
 import spock.lang.Specification
