+++ instrumentation/logback/logback-mdc-1.0/library/src/test/resources/logback.xml
@@ -18,7 +18,8 @@
 <configuration>
   <appender name="LIST" class="ch.qos.logback.core.read.ListAppender"/>
 
-  <appender name="OTEL" class="io.opentelemetry.instrumentation.logback.v1_0.OpenTelemetryAppender">
+  <appender name="OTEL"
+    class="io.opentelemetry.instrumentation.logback.mdc.v1_0.OpenTelemetryAppender">
     <appender-ref ref="LIST"/>
   </appender>
